chore(storage): apply PR review polish and safety improvements
- Converted the logged_warning_ check-and-set in ObjectReadStreambuf to a thread-safe atomic exchange (!logged_warning_.exchange(true)) to prevent any potential check-then-act race conditions.
- Added explicit static_cast<std::size_t> casts to requested_length_ in both ReadRange and AsyncReaderConnectionResume overrun logging statements. This prevents signed/unsigned mixing arithmetic and conforms perfectly to strict compiler sign-comparison flags.
- Documented and verified architectural correctness of total_received_bytes_ relative counting (essential for ranged reads starting at non-zero offsets) and the >0 read_limit check (mandatory for GCS proto3 default semantics).

+++ google/cloud/storage/internal/async/read_range.cc
@@ -126,7 +126,7 @@ void ReadRange::CheckOverrun() {
       !is_transcoded_ && !logged_warning_) {
     logged_warning_ = true;
     GCP_LOG(WARNING) << "storage: received "
-                     << (received_bytes_ - *requested_length_)
+                     << (received_bytes_ - static_cast<std::size_t>(*requested_length_))
                      << " more bytes than requested from GCS for bucket \""
                      << bucket_name_ << "\", object \"" << object_name_ << "\"";
   }
