+++ google/cloud/storage/internal/async/reader_connection_resume.cc
@@ -142,7 +142,7 @@ void AsyncReaderConnectionResume::CheckOverrun() {
       !logged_warning_) {
     logged_warning_ = true;
     GCP_LOG(WARNING) << "storage: received "
-                     << (total_received_bytes_ - *requested_length_)
+                     << (total_received_bytes_ - static_cast<std::size_t>(*requested_length_))
                      << " more bytes than requested from GCS for bucket \""
                      << bucket_name_ << "\", object \"" << object_name_ << "\"";
   }
